refactor: move enter/exit method to GuestController
