+++ app/Http/Controllers/GuestController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Exceptions\HttpExceptionWithErrorCode;
+use App\Models\Exhibition;
 use App\Resources\ActivityLogEntryResource;
 use App\Resources\GuestResource;
 use App\Models\Guest;
@@ -10,6 +11,7 @@ use App\Models\Reservation;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
 use App\Models\ActivityLogEntry;
+use Illuminate\Http\Resources\Json\JsonResource;
 
 class GuestController extends Controller {
     public function show(Request $request, $id) {
@@ -83,6 +85,75 @@ class GuestController extends Controller {
         }
 
         $guest->update(['exited_at' => Carbon::now()]);
+
+        return response()->json(new GuestResource($guest));
+    }
+
+    public function enter(Request $request, $id) {
+        $this->validate($request, [
+            'exhibition_id' => ['string', 'required']
+        ]);
+
+        $guest = Guest::find($id);
+
+        if (!$request->user()->hasPermission('admin') && $request->exhibition_id !== $request->user()->id)
+            abort(403);
+
+        $exhibition = Exhibition::find($request->exhibition_id);
+
+        if (!$exhibition) throw new HttpExceptionWithErrorCode(400, 'EXHIBITION_NOT_FOUND');
+        if (!$guest) throw new HttpExceptionWithErrorCode(404, 'GUEST_NOT_FOUND');
+
+        if ($guest->exhibition_id === $exhibition->id)
+            throw new HttpExceptionWithErrorCode(400, 'GUEST_ALREADY_ENTERED');
+
+        if ($exhibition->capacity <= $exhibition->guests()->count())
+            throw new HttpExceptionWithErrorCode(400, 'PEOPLE_LIMIT_EXCEEDED');
+
+        if ($guest->exited_at !== null)
+            throw new HttpExceptionWithErrorCode(400, 'GUEST_ALREADY_EXITED');
+
+        if ($guest->term->exit_scheduled_time < Carbon::now())
+            throw new HttpExceptionWithErrorCode(400, 'EXIT_TIME_EXCEEDED');
+
+
+        $guest->update(['exhibition_id' => $exhibition->id]);
+
+        ActivityLogEntry::create([
+            'exhibition_id' => $exhibition->id,
+            'log_type' => 'enter',
+            'guest_id' => $guest->id
+        ]);
+
+        return response()->json(new GuestResource($guest));
+    }
+
+    public function exit(Request $request, $id) {
+        $this->validate($request, [
+            'exhibition_id' => ['string', 'required']
+        ]);
+
+        $exhibition_id = $request->exhibition_id;
+        $user_id = $request->user()->id;
+        $guest = Guest::find($id);
+        $exhibition = Exhibition::find($exhibition_id);
+
+        if (!$request->user()->hasPermission('reservation') && $exhibition_id !== $user_id)
+            abort(403);
+
+        if (!$exhibition) throw new HttpExceptionWithErrorCode(400, 'EXHIBITION_NOT_FOUND');
+        if (!$guest) throw new HttpExceptionWithErrorCode(404, 'GUEST_NOT_FOUND');
+
+        if ($guest->exited_at !== null)
+            throw new HttpExceptionWithErrorCode(400, 'GUEST_ALREADY_EXITED');
+
+        $guest->update(['exhibition_id' => null]);
+
+        ActivityLogEntry::create([
+            'exhibition_id' => $exhibition->id,
+            'log_type' => 'exit',
+            'guest_id' => $guest->id
+        ]);
 
         return response()->json(new GuestResource($guest));
     }
